requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-editorjs==0.2.1
django-froala-editor==4.0.10
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
whitenoise==6.2.0
